+++ serve.ts
@@ -7,7 +7,7 @@ import {
 import { contentType } from "https://deno.land/std@0.152.0/media_types/mod.ts";
 import { serveDirWithTs } from "https://deno.land/x/ts_serve@v1.3.0/mod.ts";
 
-import type { Routing } from "./types.ts";
+import type { Routing } from "./server/types.ts";
 import { init, populationApi, prefecturesApi } from "./server/api.ts";
 import { renderToString } from "./components/Page.tsx";
